Add Docker support (#174)
* feat: add Dockerfile

Add a Dockerfile to be able to run serve the portfolio from a docker
container. The Dockerfile defines a docker image for NodeJS with gulp
installed globally (i.e. `npm install gulp --global`) that will execute
the command `gulp serve` when run.

In addition, a .dockeringore file was added which exclude NodeJS files
and potentially generated files from the docker image.

* chore: add Docker section to README

Update the README to include a section on how to use Docker to serve the 
portfolio and some basic usage.

The basic usage is not particularly useful for people used to using 
Docker, but is mainly there as a reminder for myself.

* docs: simply Docker documentation in README

* chore: improve Dockerfile and .dockerignore

* chore: run tests in Docker container

* chore: add gulp commands to interact with Docker

Add multiple gulp commands to interact with the Docker image and Docker containter:

- `gulp docker:build`: build the Docker image from the Dockerfile
- `gulp docker:rmi`: remove the Docker image from the system
- `gulp docker:start`: start a Docker container from the image
- `gulp docker:stop`: stop (and remove) the Docker container
- `gulp docker:logs`: shows the logs of the Docker container
- `gulp docker:attach`: attach a shell to the Docker container

* chore: add gulp docker:xxx commands to README

* fix: gulp install and gulp dist in Docker image